Refactor parse_locations backend into modular steps
